[configuration-validation] convert to typescript (release 1.0.0)
