--- packages/configuration-validation/.babelrc
@@ -1,11 +1,0 @@
-{
-  "presets": [
-    [
-      "env", {
-        "targets": {
-          "browsers": [ "> 0.1%", "not ie < 11" ]
-        }
-      }
-    ]
-  ]
-}
