+++ packages/configuration-validation/.eslintrc.js
@@ -1,11 +1,37 @@
 // https://eslint.org/docs/user-guide/configuring
 
 module.exports = {
-  extends: [ 'eslint:recommended' ],
-  parser: "babel-eslint",
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:jest/recommended',
+  ],
+  plugins: [
+    '@typescript-eslint',
+    'jest'
+  ],
+  parser: '@typescript-eslint/parser',
   env: {
     node: true,
     es6: true,
-    jest: true
-  }
+    'jest/globals': true
+  },
+  overrides: [
+    {
+      files: ['test/**/*.ts'],
+      rules: {
+        'jest/no-disabled-tests': 'warn',
+        'jest/no-focused-tests': 'error',
+        'jest/no-identical-title': 'error',
+        'jest/prefer-to-have-length': 'warn',
+        'jest/valid-expect': 'error',
+        '@typescript-eslint/no-var-requires': 0,
+        '@typescript-eslint/explicit-function-return-type': 0,
+        '@typescript-eslint/camelcase': 0,
+        '@typescript-eslint/no-empty-function': 0,
+        '@typescript-eslint/no-explicit-any': 0
+      }
+    }
+  ]
 };
